requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
pypdf2>=3.0.1
python-docx>=1.1.0
python-dotenv>=1.0.0
pandas>=2.1.3
numpy>=1.26.2


